requirements.txt
PyPDF2
groq
chromadb
sentence-transformers
nltk
fpdf
torch
load-dotenv
fpdf
streamlit
Pillow
uuid
tempfile